Refactored internal to pkg
