+++ cmd/treex/cmd/add_info.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/adebert/treex/internal/info"
+	"github.com/adebert/treex/pkg/info"
 	"github.com/spf13/cobra"
 )
