+++ cmd/treex/cmd/show.go
@@ -4,9 +4,9 @@ import (
 	"fmt"
 	"os"
 
-	"github.com/adebert/treex/internal/info"
-	"github.com/adebert/treex/internal/tree"
-	"github.com/adebert/treex/internal/tui"
+	"github.com/adebert/treex/pkg/info"
+	"github.com/adebert/treex/pkg/tree"
+	"github.com/adebert/treex/pkg/tui"
 	"github.com/spf13/cobra"
 )
