+++ pkg/info/.info
