+++ pkg/info/add_info_test.go
