+++ pkg/info/gen_info_test.go
