+++ pkg/info/parser.go
