+++ pkg/info/parser_test.go
