+++ pkg/tree/ignore.go
