+++ pkg/tree/ignore_test.go
