+++ pkg/tui/renderer.go
@@ -5,8 +5,8 @@ import (
 	"io"
 	"strings"
 
-	"github.com/adebert/treex/internal/tree"
-	"github.com/adebert/treex/internal/info"
+	"github.com/adebert/treex/pkg/tree"
+	"github.com/adebert/treex/pkg/info"
 )
 
 // TreeRenderer handles rendering file trees with annotations
