+++ pkg/tui/renderer_test.go
@@ -4,8 +4,8 @@ import (
 	"strings"
 	"testing"
 
-	"github.com/adebert/treex/internal/info"
-	"github.com/adebert/treex/internal/tree"
+	"github.com/adebert/treex/pkg/info"
+	"github.com/adebert/treex/pkg/tree"
 )
 
 func TestTreeRenderer_BasicTree(t *testing.T) {
