+++ pkg/tui/styled_renderer_test.go
@@ -5,8 +5,8 @@ import (
 	"testing"
 
 	"github.com/charmbracelet/lipgloss"
-	"github.com/adebert/treex/internal/info"
-	"github.com/adebert/treex/internal/tree"
+	"github.com/adebert/treex/pkg/info"
+	"github.com/adebert/treex/pkg/tree"
 )
 
 func TestStyledTreeRenderer_BasicTree(t *testing.T) {
